Avoid sending the remove request with empty objects
In previous implementation, we will call s3 func `RemoveObjects`
even if there is nothing to be deleted. In this case, the input of
this function has empty deleting `Objects`. Then the AWS somehow
update the S3 server and considers that this kind of empty input
is invalid, which triggers the following issue.

In order to avoid this issue, we need to check if the request input
is empty and there is nothing to be deleted before sending the request.

Longhorn #774

+++ s3/s3_service.go
@@ -149,18 +149,20 @@ func (s *Service) DeleteObjects(keys []string) error {
 			Key: aws.String(k),
 		}
 	}
-	quiet := true
-	params := &s3.DeleteObjectsInput{
-		Bucket: aws.String(s.Bucket),
-		Delete: &s3.Delete{
-			Objects: identifiers,
-			Quiet:   &quiet,
-		},
-	}
+	if len(identifiers) != 0 {
+		quiet := true
+		params := &s3.DeleteObjectsInput{
+			Bucket: aws.String(s.Bucket),
+			Delete: &s3.Delete{
+				Objects: identifiers,
+				Quiet:   &quiet,
+			},
+		}
 
-	resp, err := svc.DeleteObjects(params)
-	if err != nil {
-		return fmt.Errorf("failed to remove objects with param %+v: %v", params, parseAwsError(resp.String(), err))
+		resp, err := svc.DeleteObjects(params)
+		if err != nil {
+			return fmt.Errorf("failed to remove objects with param %+v: %v", params, parseAwsError(resp.String(), err))
+		}
 	}
 	return nil
 }
